requirements.txt
torch>=1.10.0
transformers>=4.15.0
python-docx>=0.8.11
spacy>=3.0.0
gradio>=3.0

# For .doc file conversion, LibreOffice must be installed on the system.
# After installing spacy, you need to download the language model:
# python -m spacy download en_core_web_sm